Add agent.py, opponent.py, README.md, samples.py and test_fullgame.py
agent.py contains the agent that will play for a specific side. This is run via test_fullgame.py which initialises my agent as black, and uses a random move opponent (in opponent.py). Samples are sample boards which can change the board state at the start of the game. The README explains the variant of chess in detail.